refactor: merge test_neutron.py into test_radiation.py; add test naming convention to AGENTS.md (#8)
test_neutron.py violated the one-to-one test_<module>.py convention
documented in AGENTS.md.  All neutron tests belong in test_radiation.py
since they test functions in radiation.py.

Also update AGENTS.md to explicitly document the test file naming rule
and its exception (cross-module regression files).

Contributed by: OpenCode (argo/claudesonnet46)
